+++ CHANGELOG.md
@@ -8,9 +8,18 @@ See (0Ver)[https://0ver.org/].
 
 ## 0.13.0
 
+### Features
+
+- Adds `io_squash` to squash several `IO` containers into one container with a tuple inside, currently works with `9` containers max at a time
+
 ### Bugfixes
 
 - Fixes that containers were not usable with `multiprocessing`
+- Changes the inheritance order, now `BaseContainer` is the first child
+
+### Misc
+
+- Now `generated` package is protected
 
 
 ## 0.12.0